+++ FixedPaletteTool/Runtime/FixedPaletteAttribute.cs
@@ -1,0 +1,16 @@
+﻿using FixedColorPaletteTool.Enums;
+using UnityEngine;
+
+namespace FixedColorPaletteTool
+{
+    public class FixedPaletteAttribute : PropertyAttribute
+    {
+        public COLOR_SELECT ColorSelect { get; }
+
+        public FixedPaletteAttribute(COLOR_SELECT colorSelect = COLOR_SELECT.DEFAULT)
+        {
+            ColorSelect = colorSelect;
+        }
+        
+    }
+}