+++ FixedPaletteTool/Runtime/FixedPaletteSettings.cs
@@ -1,0 +1,85 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
+
+namespace FixedColorPaletteTool
+{
+    public class FixedPaletteSettings : ScriptableObject
+    {
+        private const string AssetName = "FixedPaletteSettings";
+        public const string AssetPath = "Assets/Settings/" + AssetName + ".asset";
+        
+        public ColorPaletteScriptableObject selectedPalette;
+
+#if !UNITY_EDITOR
+        public static FixedPaletteSettings Instance => s_instance;
+        private static FixedPaletteSettings s_instance;
+
+        private void OnEnable()
+        {
+            // This will be called when Unity instantiates the preloaded object
+            s_instance = this;
+        }
+
+#else
+        public static FixedPaletteSettings Instance => GetOrCreate();
+        public static SerializedObject GetSerializedObject() => new(GetOrCreate());
+        
+        internal static FixedPaletteSettings GetOrCreate()
+        {
+            var settings = AssetDatabase.LoadAssetAtPath<FixedPaletteSettings>(AssetPath);
+            if (settings != null) 
+                return settings;
+            
+            settings = CreateInstance<FixedPaletteSettings>();
+            var palette = CreateInstance<ColorPaletteScriptableObject>();
+            palette.name = "Default Color Palette";
+            
+            settings.selectedPalette = palette;
+            AssetDatabase.CreateAsset(settings, AssetPath);
+            AssetDatabase.AddObjectToAsset(palette, AssetPath);
+            //Add new settings into preload assets for player runtime access
+            //-------------------------------------------------------------------//
+            var preloadedAssets = PlayerSettings.GetPreloadedAssets().ToList();
+            if(preloadedAssets.Any(x => x.GetType() == typeof(FixedPaletteSettings)))
+                throw new Exception();
+            preloadedAssets.Add(settings);
+            PlayerSettings.SetPreloadedAssets(preloadedAssets.ToArray());
+            //-------------------------------------------------------------------//
+            AssetDatabase.SaveAssets();
+            return settings;
+        }
+
+        public static ColorPaletteScriptableObject AddNewPalette()
+        {
+            const string AssetPath = "Assets/Settings/FixedPaletteSettings.asset";
+
+            var settings = GetOrCreate();
+            var palette = CreateInstance<ColorPaletteScriptableObject>();
+            palette.name = "New Color Palette";
+            palette.colors = new List<ColorData>();
+            
+            settings.selectedPalette = palette;
+            AssetDatabase.AddObjectToAsset(palette, AssetPath);
+            AssetDatabase.SaveAssets();
+            return palette;
+        }
+        public static void DeletePalette(ColorPaletteScriptableObject toDestroy)
+        {
+            var settings = GetOrCreate();
+
+            AssetDatabase.RemoveObjectFromAsset(toDestroy);
+            
+            DestroyImmediate(toDestroy);
+            
+            EditorUtility.SetDirty(settings);
+            AssetDatabase.SaveAssetIfDirty(settings);
+        }
+#endif
+    }
+}