+++ Runtime/Scripts/Utilities/Tweening/TransformTweenExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
+using System.Collections;
 using UnityEngine;
-
+using Utilities.Enums;
 using Object = UnityEngine.Object;
 
 namespace Utilities.Tweening
@@ -23,37 +24,49 @@ namespace Utilities.Tweening
         private static TweenController _tweenController;
 
 
+        #region Sync Calls
+
         #region Transform Move
 
-        public static void TweenTo(this Transform transform, Vector3 targetWorldPosition, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
+        public static void TweenTo(this Transform transform, SPACE transformSpace, Vector3 targetPosition, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
         {
             TweenController.GetTweenData(transform, TRANSFORM.MOVE)
-                .SetData(true, TRANSFORM.MOVE, transform, time, curve, onCompleted)
-                .SetTargetPosition(targetWorldPosition);
+                .SetData(transformSpace, TRANSFORM.MOVE, transform, time, curve, onCompleted)
+                .SetTargetPosition(targetPosition);
 
         }
-        public static void TweenToLocal(this Transform transform, Vector3 targetLocalPosition, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
+        
+        public static IEnumerator TweenToCoroutine(this Transform transform, SPACE transformSpace, Vector3 targetPosition, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
         {
-            TweenController.GetTweenData(transform, TRANSFORM.MOVE)
-                .SetData(false, TRANSFORM.MOVE, transform, time, curve, onCompleted)
-                .SetTargetPosition(targetLocalPosition);
+            yield return TweenController.GetTweenData(transform, TRANSFORM.MOVE)
+                .SetData(transformSpace, TRANSFORM.MOVE, transform, time, curve, null)
+                .SetTargetPosition(targetPosition)
+                .AsCoroutine();
+            
+            //We delay the callback to when the user would expect, at the end of the Coroutine
+            onCompleted?.Invoke();
         }
 
         #endregion //Transform Move
         
         #region Transform Rotate
 
-        public static void TweenTo(this Transform transform, Quaternion targetWorldRotation, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
+        public static void TweenTo(this Transform transform, SPACE transformSpace, Quaternion targetRotation, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
         {
             TweenController.GetTweenData(transform, TRANSFORM.ROTATE)
-                .SetData(true, TRANSFORM.ROTATE, transform, time, curve, onCompleted)
-                .SetTargetRotation(targetWorldRotation);
+                .SetData(transformSpace, TRANSFORM.ROTATE, transform, time, curve, onCompleted)
+                .SetTargetRotation(targetRotation);
         }
-        public static void TweenToLocal(this Transform transform, Quaternion targetLocalRotation, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
+        
+        public static IEnumerator TweenToCoroutine(this Transform transform, SPACE transformSpace, Quaternion targetRotation, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
         {
-            TweenController.GetTweenData(transform, TRANSFORM.ROTATE)
-                .SetData(false, TRANSFORM.ROTATE, transform, time, curve, onCompleted)
-                .SetTargetRotation(targetLocalRotation);
+            yield return TweenController.GetTweenData(transform, TRANSFORM.ROTATE)
+                .SetData(transformSpace, TRANSFORM.ROTATE, transform, time, curve, null)
+                .SetTargetRotation(targetRotation)
+                .AsCoroutine();
+            
+            //We delay the callback to when the user would expect, at the end of the Coroutine
+            onCompleted?.Invoke();
         }
 
         #endregion //Transform Rotate
@@ -63,12 +76,26 @@ namespace Utilities.Tweening
         public static void TweenScaleTo(this Transform transform, Vector3 targetScale, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
         {
             TweenController.GetTweenData(transform, TRANSFORM.SCALE)
-                .SetData(false, TRANSFORM.SCALE, transform, time, curve, onCompleted)
+                .SetData(SPACE.LOCAL, TRANSFORM.SCALE, transform, time, curve, onCompleted)
                 .SetTargetScale(targetScale);
+        }
+        
+        public static IEnumerator TweenScaleToCoroutine(this Transform transform, Vector3 targetScale, float time, CURVE curve = CURVE.LINEAR, Action onCompleted = null)
+        {
+            yield return TweenController.GetTweenData(transform, TRANSFORM.SCALE)
+                .SetData(SPACE.LOCAL, TRANSFORM.SCALE, transform, time, curve, null)
+                .SetTargetScale(targetScale)
+                .AsCoroutine();
+            
+            //We delay the callback to when the user would expect, at the end of the Coroutine
+            onCompleted?.Invoke();
         }
 
         #endregion //Transform Rotate
 
+        #endregion //Synchronous Calls
+
+       
         //TransformTweenExtensions Setup Functions
         //============================================================================================================//